Moved `TheComments` component to `/Posts` folder
