+++ src/pages/EntryPage.vue
@@ -110,7 +110,7 @@ import { Timestamp } from 'firebase/firestore'
 import LikesBar from 'src/components/Graphs/LikesBar.vue'
 import SharesPie from 'src/components/Graphs/SharesPie.vue'
 import ShareComponent from 'src/components/Posts/ShareComponent.vue'
-import TheComments from 'src/components/TheComments.vue'
+import TheComments from 'src/components/Posts/TheComments.vue'
 import TheHeader from 'src/components/TheHeader.vue'
 import { useCommentStore, useEntryStore, useErrorStore, useLikeStore, useShareStore, useUserStore } from 'src/stores'
 import { getStats } from 'src/utils/date'
