+++ src/pages/PromptPage.vue
@@ -122,7 +122,7 @@ import LikesBar from 'src/components/Graphs/LikesBar.vue'
 import SharesPie from 'src/components/Graphs/SharesPie.vue'
 import ShareComponent from 'src/components/Posts/ShareComponent.vue'
 import ShowcaseArt from 'src/components/Posts/ShowcaseArt.vue'
-import TheComments from 'src/components/TheComments.vue'
+import TheComments from 'src/components/Posts/TheComments.vue'
 import TheEntries from 'src/components/TheEntries.vue'
 import TheHeader from 'src/components/TheHeader.vue'
 import { useCommentStore, useEntryStore, useErrorStore, useLikeStore, usePromptStore, useShareStore, useUserStore } from 'src/stores'
